Architectures/Mixture of Agents/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alpaca_eval==0.6.6
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
datasets==3.6.0
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.18.0
fire==0.7.0
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
jiter==0.10.0
joblib==1.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.5.0
multiprocess==0.70.16
numpy==2.3.0
openai==1.90.0
packaging==25.0
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
propcache==0.3.2
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
together==1.5.17
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
